refactored api out of actions to simplify testing
